Get a local server running in the app
- this will be used for writing to the filesystem; this is mainly
to circumvent create-react-app not allowing Electron to be used
inside it (something in the Webpack configuration - I'm committing
this a couple of weeks after writing it and don't remember it,
unfortunately).
- this is also a nice separation of concerns. I'll be able to write
a clean API to handle all of these things this way.
- the server-side files were mostly copied from another project and
haven't been converted properly yet, so won't work. The server
runs, but nothing hits it yet - lots to do there!