Move and Rename a bunch of files to prepare for reworking layout
- Instead of having the sidebar as its own thing outside of the
MainWindow, I wish to convert it to an item that can fit inside a
pane like all of the other ones, and expand the MainWindow to go
all the way across the top of the app.
- in order to do this, take the non-note components out of the
sidebar, rename the sidebar to 'Notes' and its sub-components
appropriately. Leave NPCNames in this directory for now since the
place I will want it isn't build yet, and move DiceRoller into the
MainWindowHeader directory since that is where it will live.