Implement dynamic AspectSlots in MainWindow
- instead of statically declaring each AspectSlot, generate them
dynamically
- the number of them is set in state; this will be easy to make
available to the user for customization
- still trying to figure out if generateAspectSlots belongs somewhere
else; instinct is telling me it shouldn't live in the component but
I'm not sure where it should be yet
- I also think the logic for getting MainWindowBodyHeight belongs
in a custom hook. I might try that. This is not working in the UI
currently anyway, so it will be a good time to pull this out and
try a different strategy if that is appropriate.