Change terrain to location and reformat terrainTypes file
- this commit is far larger than I would like but all of this was
done together and I didn't get it building between two two big changes
- instead of using terrain throughout the app, use location instead.
A location can be a terrain or a place, and as each will have their
own color each can be a location. This sets the stage for adding
specific places to the app as well as more general terrains.
- while in here, convert terrainTypes to returning an object with
each terrain as a property on it instead of exporting each one
individually. This was creating groups of Modules instead of arrays
or objects and they were producing unexpected results when mapped
over and such. I figured it makes much more sense to use arrays
and objects as that will save headache in the long run, and as this
is a personal project (at least atm) I can afford to take it slow
and build things well instead of rushing to get an MVP out the door.
- this involves changing the way terrain is imported in many places
and also messed with a few things in Ambiences. I'm still working
on getting the right track lists to show up in each Drawer for
Environmental Tracks.
- also add a 'places' data file