Implement a context setup for getting reactive data to AspectSlots
- since AspectSlots are generated within a loop, the props they
receive do not update, since their values are inside a closure.
This is fine for receiving initial data and rendering their contents,
but for dynamic data it doesn't work.
- solve this problem by putting data they will need into a context
and passing it in that way.
- since I will need at least three contexts in this app, create a
separate component that is dedicated to wrapping the app in all of
the providers it needs, so that this logic doesn't clutter up
MainWindow unnecessarily.