Reorganize /Components directory and adjust imports (big)
- the organization in the /Components directory has been bothering
me for a while. Now that things have grown it's become apparent
that the current structure is incorrect in a few ways.
- create a /UI directory that will contain all components that are
not directly associated with an Aspect or a particular area of the
app. Inside here, sort components into Action, for interactables;
Structure, for organizing the UI; and Style, for decorating and
beautifying other cmponents.