Remove data that doesn't need to be in an open-source app
- none of this is secret, but I don't want to give the owners/
creators of any of the tracks that I've been using for my personal
game a reason to be upset by linking to their work extensively in
this project. If you are reading this and wish to use tracks from
Michael Ghelfi or Two Steps From Hell (or others), please support
them. Michael is a nice guy and doesn't ask for much, and he does
produce remarkable work.