Make a way to access the database that should work in production
- before this, I was starting up a local server in the background,
and sending API calls to that to interact with the database. But
when it was packaged for production, that server wouldn't work
correctly.
- To fix this, use Electron's IPC communication to emit and respond
to events between Electron's main and renderer processes, since the
main process can access the database. I will remove the local
server once I have everything connecting properly using the IPC
strategy.