Rebuild the server to use Electron's IPC rather than a local server
- before, I was spinning up a backend server to serve a backend, so
that I could create an API for interacting with the database. This
worked in development, but in production it would break.
- after doing some research, I discovered that Electron's main
process can access the database, and can interact with the renderer
processes (the react app is on one) through IPC, which is basically
websockets. I haven't tested it in production yet but from what
I've read I believe this will work.
- this means that the /localServer directory needed to be
rewritten; rather than receiving a request object through express,
it is now receiving a simple JS object. Accordingly, the
controllers need to be rewired as case statements which check
against a string (the route). Because this is no longer a real
server setup, rename the folder /pseudoServer. The architecture
is very similar but I don't want it to cause confusion and wasted
time if others think it's an actual server.
- update the package.json scripts and the Content Security Policy
 to reflect this.
- in order for the renderer process (the React side) to be able
to communicate with the main process, preload the ipcRenderer
API into the window object. This needs to be done because
the way create-react-app is set up, it blocks access to the
'electron' import in renderer processes - I don't understand
it entirely but it's something to do with the webpack
configuration. I realize that adding this to the window object
would usually be seen as a security risk, and it will be if
this app is ever used on the web; but it is safe enough in
an Electron app, since the only person able to access the
window object is the user themselves. (If this is not true,
they have problems far beyond the scope of this app.)
- also update StaticdataInitializer to use the new
callDatabase() function (a wrapper for IPC communication)
instead of using axios to hit the server's API.