Set up Ambience Tracks to use the HTML5 audio controls
- each track gets its own independent set of controls now, so it can
function like a mixer. It also plays tracks aloud instead of hitting the
discord bot now.
- remove my compositions from specialTracks, since they aren't going to
be able to be loaded like the other items.
- this change uses the Virtual Audio Cable idea - so instead of using a
Discord bot, we now have a second Discord user signed in and route the
computer audio through them, so we aren't depending on a bot or YouTube
anymore - now anything the computer can play can be shared in the call.

Conflicts:
src/state/StaticDataInitializer.js