Get the database hooked up and working again
- my issues since I set up my new computer have all been stemming from
the fact that I named a directory containing all of this Apps/Sites.
The '/' converts to '&:' in bash, and ':' is what npm uses to delimit
something; which is why none of my symlinks were working and 'npm
install' was so flaky.
- now that I've renamed that directory properly, everything is
continuing to work as it did before. When I got this computer, I split
a separate repo off of the original one because I thought this was all
due to compatibility issues with Apple Silicon; I worked on some other
features during that time as well. Now, I'm going to recombine the two
repositories since the compatibility issue isn't really a thing, which
will involve some fun Git work.

Conflicts:
package-lock.json
package.json
pseudoServer/server.js
src/state/StaticDataInitializer.js